n880e: revert atlas40 patches,when reverting.

+++ .myfiles/patch.sh
@@ -286,7 +286,10 @@ if [ "$mode" = "r"  -o "$lastDevice" != "$device" ]; then
      resetProject hardware/ti/wlan $branch
      resetProject bootable/recovery $branch
     
-     rm -f $basedir/.atlas40_patched
+   if [ -f $basedir/device/zte/atlas40/patches/install.sh ]; then
+       $basedir/device/zte/atlas40/patches/install.sh -r 
+       rm -f $basedir/.atlas40_patched
+   fi
 
      ### reset kernel/motorola/omap4-common
      curdir=`pwd`
@@ -456,7 +459,7 @@ elif [ "$device" = "n880e" ]; then
         updateBranch frameworks/native legaCyMod_$branch  legaCyMod $branch
    fi
 
-   if [ ! -f $basedir/.atlas40_patched ] ; then
+   if [ ! -f $basedir/.atlas40_patched ] && [ -f $basedir/device/zte/atlas40/patches/install.sh ]; then
        $basedir/device/zte/atlas40/patches/install.sh && touch $basedir/.atlas40_patched
    fi
 fi
